griptape_model_switcher/requirements.txt
python-dotenv==1.0.1
griptape[drivers-prompt-anthropic, drivers-prompt-google]==1.3.0
